created seed files for each table, added cookie-session and morgan to index.js. moved index.js to root directory so we can directly run npm run local
