+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "chalk": "^2.4.2",
+    "cookie-session": "^1.3.3",
     "dotenv": "^2.0.0",
     "express": "^4.17.1",
     "morgan": "^1.9.1",
